feat: enhance styles of landing page
