+++ apps/web/package.json
@@ -16,6 +16,7 @@
 	"dependencies": {
 		"@eventer/ui": "workspace:*",
 		"@t3-oss/env-nextjs": "^0.13.7",
+		"@tailwindcss/postcss": "^4.1.11",
 		"better-sqlite3": "^11.10.0",
 		"dotenv": "^17.0.0",
 		"drizzle-orm": "^0.44.2",
@@ -23,9 +24,11 @@
 		"mysql2": "^3.14.1",
 		"next": "^15.3.0",
 		"pg": "^8.16.0",
+		"postcss": "^8.5.6",
 		"react": "^19.1.0",
 		"react-dom": "^19.1.0",
 		"react-icons": "^5.5.0",
+		"tailwindcss": "^4.1.11",
 		"zod": "^3.25.56"
 	},
 	"devDependencies": {
