+++ apps/web/postcss.config.mjs
@@ -1,0 +1,6 @@
+const config = {
+	plugins: {
+		"@tailwindcss/postcss": {},
+	},
+};
+export default config;
