+++ apps/web/src/components/home/index.ts
@@ -1,0 +1,7 @@
+export { DemoSection } from "./DemoSection";
+export { EmailSection } from "./EmailSection";
+export { FeatureSection } from "./FeatureSection";
+export { FooterSection } from "./FooterSection";
+export { HeroSection } from "./HeroSection";
+export { ShowcaseSection } from "./ShowcaseSection";
+export { TestimonialSection } from "./TestimonialSection";
